backend/requirements.txt
# FastAPI 및 서버
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# 데이터베이스
sqlalchemy==2.0.23
alembic==1.12.1
pymysql==1.1.0
cryptography==41.0.7

# 데이터 검증
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# 유틸리티
python-dotenv==1.0.0
pandas==2.1.3
openpyxl==3.1.2

# 파일 처리
aiofiles==23.2.1
python-magic-bin==0.4.14

openai==1.12.0